Basic code to record video and break it up into files
Record video from the picamera in h264 format to a buffer and periodically write the buffer to a new file.
Files are named such that when they are ordered alphabetically, they will be in chronological order.
This means they can be combined into one vide file easily, for example `cat *.vblock > video.h264`.